Add the ability to export and import the entire the terminal.
While this approach does not feel very good, it is effective. The
internal structure of the Terminal object does not clearly define which
state is the core mutable data set. The JSON stringification and parsing
causes an object clone, which is necessary due to the continued mutation
of `this.lines` (and possibly other internal state).